Migrate the rest of the storage requests to Playwright
